Conflicts:
	README.md
	pom.xml
	src/java/htsjdk/samtools/CRAMFileReader.java
	src/java/htsjdk/samtools/CRAMIterator.java
	src/java/htsjdk/samtools/SAMFileWriterFactory.java
	src/java/htsjdk/samtools/cram/build/CramIO.java
	src/java/htsjdk/samtools/cram/index/CraiIndexer.java
	src/java/htsjdk/samtools/cram/index/CramIndex.java
	src/java/htsjdk/samtools/cram/ref/ReferenceSource.java
	src/java/htsjdk/samtools/metrics/MetricsFile.java
	src/java/htsjdk/samtools/reference/ReferenceSequenceFileFactory.java
	src/java/htsjdk/tribble/Tribble.java
	src/java/htsjdk/tribble/example/CountRecords.java
	src/java/htsjdk/tribble/index/AbstractIndex.java
	src/java/htsjdk/tribble/index/IndexFactory.java
	src/java/htsjdk/tribble/util/ParsingUtils.java
	src/tests/java/htsjdk/variant/vcf/VCFHeaderUnitTest.java

+++ src/java/htsjdk/samtools/BAMIndexer.java
@@ -281,7 +281,7 @@ public class BAMIndexer {
 
         BAMIndexer indexer = new BAMIndexer(output, reader.getFileHeader());
 
-        int totalRecords = 0;
+        long totalRecords = 0;
 
         // create and write the content
         for (SAMRecord rec : reader) {
